general setup to run plumber
