+++ tools/site_and_regional/run_tower
@@ -1,12 +1,12 @@
 #!/usr/bin/env python3
 """
 This is a just top-level skeleton script that calls
-run_neon.py.
-The original code (run_neon.py) is located under
+run_tower.py.
+The original code (run_tower.py) is located under
 python/ctsm/site_and_regional folder.
 
 For full instructions on how to run the code and different options,
-please check python/ctsm/site_and_regional/run_neon.py file.
+please check python/ctsm/site_and_regional/run_tower.py file.
 
 This script first creates and builds a generic base case.
 Next, it will clone the base_case for different neon sites and run
@@ -25,7 +25,7 @@ This script will do the following:
 
 ----------------------------------------------------------------
 To see all available options for running tower sites:
-    ./run_neon --help
+    ./run_tower --help
 ----------------------------------------------------------------
 Instructions for running using conda python environments:
 ../../py_env_create
@@ -42,7 +42,7 @@ _CTSM_PYTHON = os.path.join(
 sys.path.insert(1, _CTSM_PYTHON)
 
 # pylint: disable=import-error, wrong-import-position
-from ctsm.site_and_regional.run_neon import main
+from ctsm.site_and_regional.run_tower import main
 
 if __name__ == "__main__":
     main(__doc__)
